requirements.txt
tensorflow-gpu==2.3.0
tensorflow-addons
tensorflow-graphics-gpu
numpy
scipy
opencv-python
plyfile
pandas
matplotlib
PyYAML
scikit-image
deprecated
silence_tensorflow
